continuacion de estandarizacion, tabla y botones de accion
