+++ src/Modules/Donors/Styles/DonorForm.css
@@ -744,6 +744,7 @@
   border-color: rgba(255,255,255,0.2);
   background: rgba(255,255,255,0.06);
 }
+[data-theme="dark"] .donor-form__donation-type { color: #d1d1d6; }
 [data-theme="dark"] .donor-form__donation-desc { color: #d1d1d6; }
 [data-theme="dark"] .donor-form__donation-date { color: #6e6e73; }
 [data-theme="dark"] .donor-form__inline-edit {
