+++ src/Modules/Users/Styles/UsersPage.css
@@ -84,6 +84,21 @@
   flex-wrap: wrap;
 }
 
+/* ── Role filter group ── */
+.users-page__filter-group {
+  display: flex;
+  align-items: center;
+  gap: 0.375rem;
+  flex-shrink: 0;
+}
+
+.users-page__filter-label {
+  font-size: 0.8125rem;
+  font-weight: 500;
+  color: var(--t2);
+  white-space: nowrap;
+}
+
 /* ── Search ── */
 .users-page__search-wrapper {
   position: relative;
