+++ src/Modules/Volunteers/Styles/VolunteersTable.css
@@ -1,59 +1,83 @@
 .volunteers-table {
   width: 100%;
   border-collapse: collapse;
-  background: white;
-  border-radius: 0.75rem;
+  background: var(--bg-card);
+  border-radius: var(--r, 12px);
   overflow: hidden;
-  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1);
+  border: 1px solid var(--border-strong, #d1d5db);
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.06);
   margin-top: 1rem;
 }
 
 .volunteers-table th:nth-child(1),
 .volunteers-table td:nth-child(1) {
-  max-width: 150px; 
+  max-width: 150px;
 }
 
 .volunteers-table th:nth-child(2),
 .volunteers-table td:nth-child(2) {
-  max-width: 180px; 
+  max-width: 180px;
+}
+
+.volunteers-table td:nth-child(2) {
+  color: var(--t2);
 }
 
 .volunteers-table th:nth-child(3),
 .volunteers-table td:nth-child(3) {
-  max-width: 150px; 
+  max-width: 150px;
 }
 
-.volunteers-table th,
-.volunteers-table td {
-  padding: 0.75rem 1rem;
+.volunteers-table thead tr {
+  border-bottom: 1px solid var(--border-strong, #d1d5db);
+}
+
+.volunteers-table th {
+  padding: 0.625rem 0.875rem;
   text-align: left;
-  font-size: 0.9rem;
-  color: #374151;
-  border-bottom: 1px solid #e5e7eb;
+  font-size: 0.75rem;
+  font-weight: 600;
+  color: var(--t2);
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+  white-space: nowrap;
+  background: var(--bg);
+}
+
+.volunteers-table td {
+  padding: 0.75rem 0.875rem;
+  text-align: left;
+  font-size: 0.8125rem;
+  color: var(--t1);
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
   vertical-align: middle;
 }
 
-.volunteers-table th {
-  background-color: #f9fafb;
-  font-weight: 600;
-  font-size: 0.95rem;
-  color: #1f2937;
+.volunteers-table tbody tr {
+  border-bottom: 1px solid var(--border-strong, #d1d5db);
+  transition: background 0.1s;
+}
+
+.volunteers-table tbody tr:last-child {
+  border-bottom: none;
 }
 
 .volunteers-table tbody tr:hover {
-  background-color: #f3f4f6;
-}
-
-.volunteers-table tbody tr:last-child td {
-  border-bottom: none;
+  background: var(--bg, #f3f4f6);
 }
 
 .volunteers-table .table-actions {
   display: flex;
-  gap: 0.5rem;
+  gap: 0.375rem;
+  align-items: center;
+}
+
+.volunteers-table .table-actions .edit-volunteer-button {
+  width: auto;
+  padding: 0.4375rem 0.75rem;
+  font-size: 0.8125rem;
 }
 
 .volunteers-table .table-actions .view-icon {
@@ -66,19 +90,22 @@
   height: 1rem;
 }
 
+/* Ver Detalles */
 .volunteers-table .table-actions button.view {
-  display: flex;
+  display: inline-flex;
   align-items: center;
   justify-content: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1rem;
+  gap: 0.375rem;
+  padding: 0.4375rem 0.75rem;
   background-color: #f8fafc;
   border: 1px solid #e2e8f0;
   border-radius: 0.5rem;
   color: #475569;
+  font-size: 0.8125rem;
   font-weight: 500;
   cursor: pointer;
   transition: all 0.2s ease-in-out;
+  white-space: nowrap;
 }
 
 .volunteers-table .table-actions button.view:hover {
@@ -87,7 +114,9 @@
   color: #334155;
 }
 
+/* Estado pill */
 .volunteers-table__status {
+  display: inline-block;
   padding: 0.25rem 0.75rem;
   border-radius: 9999px;
   font-size: 0.75rem;
@@ -97,7 +126,7 @@
 
 .volunteers-table__status--active {
   background-color: #d1fae5;
-  color: #059669;
+  color: #065f46;
 }
 
 .volunteers-table__status--inactive {
@@ -105,17 +134,20 @@
   color: #dc2626;
 }
 
+/* Activar/Inactivar */
 .volunteers-table .table-actions button.toggle {
-  display: flex;
+  display: inline-flex;
   align-items: center;
   justify-content: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1rem;
+  gap: 0.375rem;
+  padding: 0.4375rem 0.5rem;
   border-radius: 0.5rem;
+  font-size: 0.8125rem;
   font-weight: 500;
   cursor: pointer;
   transition: all 0.2s ease-in-out;
   border: 1px solid transparent;
+  white-space: nowrap;
 }
 
 .volunteers-table .table-actions button.toggle.active {
@@ -140,7 +172,7 @@
   border-color: #34d399;
 }
 
-/* Estados de carga */
+/* Estado de carga */
 .volunteers-table .table-actions button.loading {
   opacity: 0.7;
   cursor: not-allowed;
@@ -148,8 +180,8 @@
 
 .volunteers-table .table-actions button.loading::after {
   content: '';
-  width: 16px;
-  height: 16px;
+  width: 14px;
+  height: 14px;
   border: 2px solid transparent;
   border-top: 2px solid currentColor;
   border-radius: 50%;
@@ -161,68 +193,61 @@
   100% { transform: rotate(360deg); }
 }
 
-/* Responsive */
+/* ── Responsive ── */
 @media (max-width: 1024px) {
-  .volunteers-table .table-actions {
-    flex-wrap: wrap;
-  }
-  
-  .volunteers-table .table-actions button {
-    flex: 1;
-    min-width: 100px;
+  .volunteers-table th,
+  .volunteers-table td {
+    padding: 0.5rem 0.75rem;
+    font-size: 0.75rem;
   }
 }
 
 @media (max-width: 768px) {
+  .volunteers-table th,
+  .volunteers-table td {
+    padding: 0.5rem 0.625rem;
+    font-size: 0.75rem;
+  }
+
   .volunteers-table .table-actions {
     flex-direction: column;
-    gap: 0.5rem;
+    gap: 0.375rem;
   }
 
   .volunteers-table .table-actions button {
     width: 100%;
     justify-content: center;
+    padding: 0.5rem;
+    font-size: 0.75rem;
+    min-height: 44px;
   }
 }
 
 /* ── Dark mode ── */
-[data-theme="dark"] .volunteers-table {
-  background: #2c2c2e;
-  box-shadow: 0 4px 6px -1px rgba(0,0,0,0.3);
-}
-[data-theme="dark"] .volunteers-table th,
-[data-theme="dark"] .volunteers-table td {
-  color: #8e8e93;
-  border-bottom-color: rgba(255,255,255,0.07);
-}
-[data-theme="dark"] .volunteers-table th {
-  background-color: #1c1c1e;
-  color: #f5f5f7;
-}
-[data-theme="dark"] .volunteers-table tbody tr:hover { background-color: #363638; }
 [data-theme="dark"] .volunteers-table .table-actions button.view {
-  background-color: #363638;
+  background-color: #3a3a3c;
   border-color: rgba(255,255,255,0.1);
-  color: #8e8e93;
+  color: #d1d1d6;
 }
 [data-theme="dark"] .volunteers-table .table-actions button.view:hover {
-  background-color: #444446;
+  background-color: #48484a;
+  color: #f5f5f7;
 }
 [data-theme="dark"] .volunteers-table .table-actions button.toggle.active {
   background-color: rgba(220,38,38,0.15);
-  border-color: rgba(220,38,38,0.3);
-  color: #fca5a5;
+  border-color: rgba(248,113,113,0.3);
+  color: #f87171;
 }
 [data-theme="dark"] .volunteers-table .table-actions button.toggle.inactive {
-  background-color: rgba(5,150,105,0.15);
-  border-color: rgba(5,150,105,0.3);
-  color: #6ee7b7;
+  background-color: rgba(6,95,70,0.2);
+  border-color: rgba(52,211,153,0.3);
+  color: #34d399;
 }
 [data-theme="dark"] .volunteers-table__status--active {
-  background-color: rgba(22, 101, 52, 0.25);
+  background-color: rgba(6,95,70,0.2);
   color: #4ade80;
 }
 [data-theme="dark"] .volunteers-table__status--inactive {
-  background-color: rgba(153, 27, 27, 0.25);
+  background-color: rgba(220,38,38,0.15);
   color: #f87171;
-}
+}
